First commit of stuff from Tremolo. C-only changes to start with.
Mostly squashing of warnings with some explicit casts, and removing some
variables that are never used.

Optimisations to render_line.

Various checks for (m)alloc failures and out of range inputs.

Change to use explicit dec_buf rather than repeated calls to alloca.

Change to use dec_method/switch rather than nested if.

Reorder of contents of codebook struct in preparation for ARM code.



git-svn-id: https://svn.xiph.org/branches/lowmem-branch/Tremolo@17217 0101bb08-14d6-0310-b084-bc0e0c8e3800

+++ dsp.c
@@ -263,10 +263,10 @@ int vorbis_dsp_synthesis(vorbis_dsp_state *vd,ogg_packet *op,int decodep){
 	  /* granulepos could be -1 due to a seek, but that would result
 	     in a long coun t, not short count */
 	  
-	  vd->out_end-=vd->sample_count-vd->granulepos;
+	  vd->out_end-=(int)(vd->sample_count-vd->granulepos);
 	}else{
 	  /* trim the beginning */
-	  vd->out_begin+=vd->sample_count-vd->granulepos;
+	  vd->out_begin+=(int)(vd->sample_count-vd->granulepos);
 	  if(vd->out_begin>vd->out_end)
 	    vd->out_begin=vd->out_end;
 	}
@@ -280,7 +280,7 @@ int vorbis_dsp_synthesis(vorbis_dsp_state *vd,ogg_packet *op,int decodep){
     if(op->granulepos!=-1 && vd->granulepos!=op->granulepos){
       
       if(vd->granulepos>op->granulepos){
-	long extra=vd->granulepos-op->granulepos;
+	long extra=(long)(vd->granulepos-op->granulepos);
 	
 	if(extra)
 	  if(op->e_o_s){
